Estabilidad gateway: httpd stack, auto_cycle async y WiFi STA
- http_server_start: stack_size=8192 para evitar overflow del task httpd
  con los buffers grandes (buf[512]+resp[2048]) del handler de comandos.
- kinco_auto_cycle: ahora corre en tarea dedicada (auto_cycle_task) en vez
  de bloquear el unico task del servidor HTTP hasta ~120 s. Nuevo comando
  kinco_cycle_status para sondear progreso; la UI actualiza el contador en
  vivo durante el ciclo en lugar de quedar congelada hasta el final.
- bridge_rs485_send: corrige comparacion signed/unsigned (written vs len).
- main: comentarios reflejan modo STA puro (AP deshabilitado), conecta a
  la red del laboratorio NS-LAB con credenciales embebidas.

+++ main/bridge_rs485.cpp
@@ -552,7 +552,7 @@ esp_err_t bridge_rs485_send(const uint8_t *data, size_t len)
     int written = uart_write_bytes((uart_port_t)s_bridge_config.rs485_uart_num,
                                    data, len);
     xSemaphoreGive(s_uart_mutex);
-    return (written == len) ? ESP_OK : ESP_FAIL;
+    return (written == (int)len) ? ESP_OK : ESP_FAIL;
 }
 
 esp_err_t bridge_rs485_transact(const uint8_t *request, size_t request_len,
